requirements.txt
openai
python-dotenv
pydantic
openai-agents
PyGithub
GitPython
atlassian-python-api